fix structure and descriptions
