+++ .env.example
@@ -1,2 +1,4 @@
 NEXT_PUBLIC_SANITY_PROJECT_ID=
-# NEXT_PUBLIC_SANITY_DATASET=
+NEXT_PUBLIC_SANITY_API_VERSION=
+NEXT_PUBLIC_SANITY_DATASET=
+NEXT_PUBLIC_SANITY_CDN=
