+++ admin/requests/queries/about.ts
@@ -1,0 +1,5 @@
+import { groq } from "next-sanity";
+
+export const aboutQuery = groq`
+   *[_type == "about"]{_id, image}
+`;
