+++ admin/requests/queries/contacts.ts
@@ -1,0 +1,5 @@
+import { groq } from "next-sanity";
+
+export const contactsQuery = groq`
+   *[_type == "contacts"]{_id, phone, email, instagram, facebook}
+`;
