+++ admin/requests/queries/gallery.ts
@@ -1,0 +1,5 @@
+// import { groq } from "next-sanity";
+
+// export const galleryQuery = groq`
+//    *[_type == "gallery"]{_id, images[]->}
+// `;
