+++ admin/requests/queries/reviews.ts
@@ -1,0 +1,5 @@
+// import { groq } from "next-sanity";
+
+// export const reviewsQuery = groq`
+//    *[_type == "reviews"][0]{_id, images}
+// `;
