+++ admin/requests/queries/services.ts
@@ -1,0 +1,5 @@
+// import { groq } from "next-sanity";
+
+// export const servicesQuery = groq`
+//    *[_type == "services"][0]{_id, images}
+// `;
