+++ admin/schemas/about.ts
@@ -9,7 +9,7 @@ const about = {
   fields: [
     {
       name: "image",
-      title: "Distributor image",
+      title: "Зображення для секції 'Про нас'",
       type: "accessibleImage",
       description:
         "Додайте зображення розмірами не менше 592 пікселів шириною та 645 пікселів висотою",
