+++ src/data/types.ts
@@ -1,0 +1,8 @@
+export type ImgType = {
+  image: {
+    image: string;
+    alt: string;
+  };
+};
+
+export type AboutResponse = ImgType;
